requirements.txt
numpy>=2.0
pandas>=2.2
matplotlib>=3.9
scikit-learn>=1.5
python-chess>=1.999
pypdf>=6.0
